(feat) Factor out cert handling and add C binding
